# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse_null_": {
				"properties": {
					"error": {
						"type": "string"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object"
			},
			"CreateNewUserRequest": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"handle": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"email",
					"handle"
				],
				"type": "object"
			},
			"GetUserIdFromTokenResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id"
				],
				"type": "object"
			},
			"ApiResponse_GetUserIdFromTokenResponse_": {
				"properties": {
					"error": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/GetUserIdFromTokenResponse"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object"
			},
			"GetUserRoleFromTokenResponse": {
				"properties": {
					"role": {
						"type": "string"
					}
				},
				"required": [
					"role"
				],
				"type": "object"
			},
			"ApiResponse_GetUserRoleFromTokenResponse_": {
				"properties": {
					"error": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/GetUserRoleFromTokenResponse"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object"
			},
			"Role": {
				"enum": [
					"ADMIN",
					"USER",
					"ORGANIZER",
					"PROBLEM_SETTER"
				],
				"type": "string"
			},
			"Problem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"title": {
						"type": "string"
					},
					"statement": {
						"type": "string"
					},
					"difficulty": {
						"type": "number",
						"format": "double"
					},
					"timeLimit": {
						"type": "number",
						"format": "double"
					},
					"memoryLimit": {
						"type": "number",
						"format": "double"
					},
					"inputFormat": {
						"type": "string"
					},
					"outputFormat": {
						"type": "string"
					},
					"solutionText": {
						"type": "string"
					},
					"owner": {
						"$ref": "#/components/schemas/User"
					},
					"submissions": {
						"items": {
							"$ref": "#/components/schemas/Submission"
						},
						"type": "array"
					},
					"testcases": {
						"items": {
							"$ref": "#/components/schemas/Testcase"
						},
						"type": "array"
					},
					"isPublished": {
						"type": "boolean"
					},
					"associatedContests": {
						"items": {
							"$ref": "#/components/schemas/ProblemInContest"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"title",
					"statement",
					"difficulty",
					"timeLimit",
					"memoryLimit",
					"inputFormat",
					"outputFormat",
					"solutionText",
					"owner",
					"submissions",
					"testcases",
					"isPublished",
					"associatedContests"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"handle": {
						"type": "string"
					},
					"avatar_image": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"isBan": {
						"type": "boolean"
					},
					"lastTimeChangeHandle": {
						"type": "number",
						"format": "double"
					},
					"lastTimeChangeImage": {
						"type": "number",
						"format": "double"
					},
					"country": {
						"type": "string"
					},
					"problems": {
						"items": {
							"$ref": "#/components/schemas/Problem"
						},
						"type": "array"
					},
					"submissions": {
						"items": {
							"$ref": "#/components/schemas/Submission"
						},
						"type": "array"
					},
					"organizedContests": {
						"items": {
							"$ref": "#/components/schemas/Contest"
						},
						"type": "array"
					},
					"participatedContest": {
						"items": {
							"$ref": "#/components/schemas/ContestParticipation"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"handle",
					"avatar_image",
					"bio",
					"email",
					"password",
					"role",
					"isBan",
					"lastTimeChangeHandle",
					"lastTimeChangeImage",
					"country",
					"problems",
					"submissions",
					"organizedContests",
					"participatedContest"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Contest": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"ruleText": {
						"type": "string"
					},
					"startTime": {
						"type": "number",
						"format": "double"
					},
					"endTime": {
						"type": "number",
						"format": "double"
					},
					"scoringRule": {
						"type": "string"
					},
					"isPlagiarismCheckEnabled": {
						"type": "boolean"
					},
					"isPublished": {
						"type": "boolean"
					},
					"organizer": {
						"$ref": "#/components/schemas/User"
					},
					"participations": {
						"items": {
							"$ref": "#/components/schemas/ContestParticipation"
						},
						"type": "array"
					},
					"problemsInContest": {
						"items": {
							"$ref": "#/components/schemas/ProblemInContest"
						},
						"type": "array"
					},
					"submissions": {
						"items": {
							"$ref": "#/components/schemas/Submission"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"title",
					"description",
					"ruleText",
					"startTime",
					"endTime",
					"scoringRule",
					"isPlagiarismCheckEnabled",
					"isPublished",
					"organizer",
					"participations",
					"problemsInContest",
					"submissions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ContestParticipation": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"contestId": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"contest": {
						"$ref": "#/components/schemas/Contest"
					},
					"score": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"userId",
					"contestId",
					"user",
					"contest",
					"score"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProblemInContest": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"problem": {
						"$ref": "#/components/schemas/Problem"
					},
					"contest": {
						"$ref": "#/components/schemas/Contest"
					},
					"score": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"problem",
					"contest",
					"score"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Submission": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"owner": {
						"$ref": "#/components/schemas/User"
					},
					"problem": {
						"$ref": "#/components/schemas/Problem"
					},
					"contest": {
						"$ref": "#/components/schemas/Contest"
					},
					"code": {
						"type": "string"
					},
					"language": {
						"$ref": "#/components/schemas/LANGUAGE"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/SubmissionResult"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"owner",
					"problem",
					"contest",
					"code",
					"language",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Testcase": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"problem": {
						"$ref": "#/components/schemas/Problem"
					},
					"input": {
						"type": "string"
					},
					"output": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"problem",
					"input",
					"output"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Pick_User.Exclude_keyofUser.password__": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"handle": {
						"type": "string"
					},
					"avatar_image": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"isBan": {
						"type": "boolean"
					},
					"lastTimeChangeHandle": {
						"type": "number",
						"format": "double"
					},
					"lastTimeChangeImage": {
						"type": "number",
						"format": "double"
					},
					"country": {
						"type": "string"
					},
					"problems": {
						"items": {
							"$ref": "#/components/schemas/Problem"
						},
						"type": "array"
					},
					"submissions": {
						"items": {
							"$ref": "#/components/schemas/Submission"
						},
						"type": "array"
					},
					"organizedContests": {
						"items": {
							"$ref": "#/components/schemas/Contest"
						},
						"type": "array"
					},
					"participatedContest": {
						"items": {
							"$ref": "#/components/schemas/ContestParticipation"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"handle",
					"avatar_image",
					"bio",
					"email",
					"role",
					"isBan",
					"lastTimeChangeHandle",
					"lastTimeChangeImage",
					"country",
					"problems",
					"submissions",
					"organizedContests",
					"participatedContest",
					"createdAt",
					"updatedAt",
					"deletedAt"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_User.password_": {
				"$ref": "#/components/schemas/Pick_User.Exclude_keyofUser.password__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"GetUserResponse": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Omit_User.password_"
					}
				],
				"nullable": true
			},
			"ApiResponse_GetUserResponse_": {
				"properties": {
					"error": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/GetUserResponse"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object"
			},
			"UpdateUserRequest": {
				"properties": {
					"country": {
						"type": "string"
					},
					"avatar_image": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"handle": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"LoginResponse": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object"
			},
			"ApiResponse_LoginResponse_": {
				"properties": {
					"error": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/LoginResponse"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object"
			},
			"LoginRequest": {
				"properties": {
					"password": {
						"type": "string"
					},
					"handle": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"password"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "web-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/user": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateNewUserRequest"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "x-access-token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserRequest"
							}
						}
					}
				}
			}
		},
		"/user/id": {
			"get": {
				"operationId": "GetUserIdFromToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_GetUserIdFromTokenResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"USER"
						]
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "x-access-token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/role": {
			"get": {
				"operationId": "GetUserRoleFromToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_GetUserRoleFromTokenResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"USER"
						]
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "x-access-token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_GetUserResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"USER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user/handle/{handle}": {
			"get": {
				"operationId": "GetUserByHandle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_GetUserResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "handle",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/id/{id}": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user/admin/{id}": {
			"patch": {
				"operationId": "AddAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_LoginResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				}
			}
		},
		"/submission": {
			"post": {
				"operationId": "Submit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"message": {
													"type": "string"
												},
												"submissionId": {}
											},
											"required": [
												"message",
												"submissionId"
											],
											"type": "object"
										},
										{
											"properties": {
												"message": {},
												"submissionId": {},
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "x-access-token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"code": {
										"type": "string"
									},
									"contestId": {
										"type": "number",
										"format": "double"
									},
									"problemId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"code",
									"problemId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/submission/execute": {
			"post": {
				"operationId": "ExecuteSubmission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"message": {},
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"submissionId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"submissionId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/submission/{id}": {
			"get": {
				"operationId": "GetSubmissionById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"data": {},
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {},
												"data": {}
											},
											"required": [
												"data"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/submission/execute?code={code}&stdin={stdin}": {
			"get": {
				"operationId": "ExecuteCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "stdin",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}